refactor: app drawer
